fix a few warning in dicom toolbox

+++ gadgets/dicom/dicom_ismrmrd_utility.cpp
@@ -668,7 +668,7 @@ namespace Gadgetron
             // Series Number
             // Only write a number if the image_series_index is positive and non-zero
             key.set(0x0020, 0x0011);
-            ACE_OS::snprintf(buf, BUFSIZE, "%ld", m1.image_series_index);
+            ACE_OS::snprintf(buf, BUFSIZE, "%ld", (long int)m1.image_series_index);
             write_dcm_string(dataset, key, buf);
 
             // Image Number
@@ -926,7 +926,7 @@ namespace Gadgetron
                 double v = attrib.as_double(GADGETRON_IMAGE_WINDOWCENTER, 0);
 
                 key.set(0x0028, 0x1050);
-                ACE_OS::snprintf(buf, BUFSIZE, "%d", v);
+                ACE_OS::snprintf(buf, BUFSIZE, "%d", (int)v);
                 write_dcm_string(dataset, key, buf);
             }
 
@@ -935,7 +935,7 @@ namespace Gadgetron
                 double v = attrib.as_double(GADGETRON_IMAGE_WINDOWWIDTH, 0);
 
                 key.set(0x0028, 0x1051);
-                ACE_OS::snprintf(buf, BUFSIZE, "%d", v);
+                ACE_OS::snprintf(buf, BUFSIZE, "%d", (int)v);
                 write_dcm_string(dataset, key, buf);
             }
         }
